Moved Rack svg code (nanosvg to nanovg glue) into separate file svg.cpp
